+++ particle_filter/modularized_particle_filter/include/modularized_particle_filter/prediction/predictor.hpp
@@ -16,6 +16,7 @@
 #define MODULARIZED_PARTICLE_FILTER__PREDICTION__PREDICTOR_HPP_
 
 #include "modularized_particle_filter/common/visualize.hpp"
+#include "modularized_particle_filter/prediction/experimental/suspension_adaptor.hpp"
 #include "modularized_particle_filter/prediction/resampler.hpp"
 
 #include <rclcpp/rclcpp.hpp>
@@ -25,7 +26,6 @@
 #include <geometry_msgs/msg/twist_stamped.hpp>
 #include <geometry_msgs/msg/twist_with_covariance_stamped.hpp>
 #include <modularized_particle_filter_msgs/msg/particle_array.hpp>
-#include <sensor_msgs/msg/image.hpp>
 #include <std_msgs/msg/float32.hpp>
 
 #include <tf2_ros/transform_broadcaster.h>
@@ -43,7 +43,6 @@ public:
   using TwistCovStamped = geometry_msgs::msg::TwistWithCovarianceStamped;
   using TwistStamped = geometry_msgs::msg::TwistStamped;
   using OptParticleArray = std::optional<ParticleArray>;
-  using Image = sensor_msgs::msg::Image;
 
   Predictor();
 
@@ -93,23 +92,6 @@ private:
   void update_with_dynamic_noise(ParticleArray & particle_array, const TwistCovStamped & twist);
 
   void publish_mean_pose(const geometry_msgs::msg::Pose & mean_pose, const rclcpp::Time & stamp);
-
-  struct SwapModeAdaptor
-  {
-    SwapModeAdaptor(rclcpp::Node * node);
-    std::optional<rclcpp::Time> stamp_opt_{std::nullopt};
-    std::optional<PoseCovStamped> init_pose_opt_{std::nullopt};
-    rclcpp::Subscription<PoseCovStamped>::SharedPtr sub_pose_;
-    rclcpp::Subscription<Image>::SharedPtr sub_image_;
-    rclcpp::Clock::SharedPtr clock_;
-
-    bool state_is_active;
-    bool state_is_activating;
-
-    bool should_keep_update();
-    bool should_call_initialize();
-    PoseCovStamped init_pose() { return init_pose_opt_.value(); }
-  };
 
   std::unique_ptr<SwapModeAdaptor> swap_mode_adaptor_{nullptr};
 };
