+++ particle_filter/modularized_particle_filter/src/prediction/predictor.cpp
@@ -372,51 +372,4 @@ void Predictor::publish_mean_pose(
   }
 }
 
-Predictor::SwapModeAdaptor::SwapModeAdaptor(rclcpp::Node * node)
-{
-  auto on_ekf_pose = [this](const PoseCovStamped & pose) -> void { init_pose_opt_ = pose; };
-  auto on_image = [this](const Image & msg) -> void {
-    stamp_opt_ = rclcpp::Time(msg.header.stamp);
-  };
-
-  sub_image_ = node->create_subscription<Image>("image", 1, on_image);
-  sub_pose_ = node->create_subscription<PoseCovStamped>("pose_cov", 1, on_ekf_pose);
-  clock_ = node->get_clock();
-
-  state_is_active = false;
-  state_is_activating = false;
-}
-
-bool Predictor::SwapModeAdaptor::should_keep_update()
-{
-  if (!stamp_opt_.has_value()) {
-    RCLCPP_INFO_STREAM(rclcpp::get_logger("swap_adaptor"), "system should stop");
-    state_is_active = false;
-    return false;
-  }
-
-  double dt = (clock_->now() - stamp_opt_.value()).seconds();
-  if (dt > 3) {
-    RCLCPP_INFO_STREAM(rclcpp::get_logger("swap_adaptor"), "system should stop");
-    state_is_active = false;
-    return false;
-  } else {
-    RCLCPP_INFO_STREAM(rclcpp::get_logger("swap_adaptor"), "system should keep estimation");
-    state_is_activating = true;
-    return true;
-  }
-}
-
-bool Predictor::SwapModeAdaptor::should_call_initialize()
-{
-  if (state_is_activating && init_pose_opt_.has_value()) {
-    if (!state_is_active) {
-      state_is_active = true;
-      state_is_activating = false;
-      return true;
-    }
-  }
-  return false;
-}
-
 }  // namespace pcdless::modularized_particle_filter
